+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Check for valid data types in service fields
 - added more valid data types
 
+### Changed
+
+- Naming of entities
+- Modularized database tables
+
 ## [0.0.7] - 2019-04-24
 
 ### Added