+++ CHANGELOG.md
@@ -6,9 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.0.8] - 2019-04-30
+
 ### Added
 
 - Check for valid data types in service fields
+- added more valid data types
+
+### Changed
+
+- Naming of entities
+- Modularized database tables
 
 ## [0.0.7] - 2019-04-24
 
@@ -67,7 +75,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - base app
 
-[Unreleased]: https://github.com/siewert87/aaas-api/compare/v0.0.7...HEAD
+[Unreleased]: https://github.com/siewert87/aaas-api/compare/v0.0.8...HEAD
+[0.0.8]: https://github.com/siewert87/aaas-api/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/siewert87/aaas-api/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/siewert87/aaas-api/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/siewert87/aaas-api/compare/v0.0.4...v0.0.5